WIP: basic tab completion
This first implementation is to be quite naïve, only completing the
files in the current directory. Currently it does _not_ work, and
crashes frequently.

Once the crashes have been ironed out, a good next step would be to be
more aware when completing. For instance, writing `~/.con<Tab>` should
complete `~/.config`. Currently it would complete to `~/.con<first file
in currend dir>`.